pyproject.toml
[project]
name = "pidili"
version = "0.1.0"
description = "Declarative display library for smartscreens based on Pillow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest-benchmark>=5.1.0",
    "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pidili"]
